[message] copy on write based on read-only flag
This commit adds a new flag to Message metadata to indicate whether or
not the message data is read-only. With this information, the upper
layer can choose clone a message only when it's read-only and
modification is going to be performed.

The typical case is the UDP payload handler removes the IP and UDP
headers so that CoAP message can have enough space for caching basic
CoAP metadata. When a message doesn't need to be forwarded, the UDP
dispatcher can safely performs the change without cloning the message.

This commit also simplifies the filter logic based on the result of
Ip6::HandlePayload. By trying to handle the payload before calling the
callback, the result tells whether the message is an Thread traffic or
not. On OT_ERROR_NO_ROUTE, it indicates OpenThread's built-in stack
cannot handle the message so deliver it to Host when possible, otherwise
not.

In typical RCP(PLATFORM_UDP is on) case, ip6 receive filter should always
be off, this commit also conditions the flag out.